@@ -2,65 +2,44 @@
   "name": "@modular-rocks/slimfast-turbo",
   "version": "0.1.1-development",
   "description": "Slimfast super-repo",
-  "main": "./dist/index.js",
-  "exports": {
-    ".": {
-      "default": "./dist/index.js",
-      "types": "./dist/declarations.d.ts"
-    },
-    "./package.json": "./package.json"
-  },
-  "types": "./dist/declarations.d.ts",
+  "private": true,
   "engines": {
-    "node": "16.x"
+    "node": ">=16.x"
   },
-  "files": [
-    "dist"
-  ],
   "scripts": {
-    "prebuild": "yarn lint",
-    "build": "tsc -p tsconfig.prod.json",
-    "prepublishOnly": "npm run build",
-    "test": "jest",
-    "semantic-release": "semantic-release",
-    "travis-deploy-once": "travis-deploy-once",
-    "lint": "yarn eslint src --ext .ts --fix"
+    "build": "turbo run build",
+    "clean": "turbo daemon clean",
+    "build:f": "turbo run build --force",
+    "test": "turbo run test",
+    "test:f": "turbo run test --force",
+    "dev": "turbo run dev --parallel",
+    "lint": "turbo run lint",
+    "format": "prettier --write \"**/*.{ts,tsx,md}\""
   },
   "devDependencies": {
-    "@babel/types": "7.22.5",
-    "@types/jest": "^29.4.0",
-    "@typescript-eslint/eslint-plugin": "^5.59.9",
-    "@typescript-eslint/parser": "^5.59.9",
-    "babel-jest": "^26.0.1",
-    "eslint": "^7.0.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-prettier": "^4.2.1",
+    "@types/jest": "29.4.0",
+    "@typescript-eslint/eslint-plugin": "5.59.9",
+    "@typescript-eslint/parser": "5.59.9",
+    "babel-jest": "26.0.1",
+    "eslint": "7.0.0",
+    "eslint-config-airbnb-base": "15.0.0",
+    "eslint-config-airbnb-typescript": "17.0.0",
+    "eslint-config-prettier": "8.8.0",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-prettier": "4.2.1",
     "jest": "29.4.1",
-    "path-browserify": "^1.0.1",
-    "prettier": "^2.8.8",
-    "ts-jest": "^29.0.5",
-    "ts-loader": "^9.4.2",
-    "typescript": "^4.9.5"
+    "prettier": "3.0.0",
+    "ts-jest": "29.1.1",
+    "ts-loader": "9.4.2",
+    "turbo": "1.10.12",
+    "typescript": "5.1.6"
   },
+  "packageManager": "pnpm@8.6.6",
   "author": "Alex Dollery (DolNpm/DolGit)",
   "homepage": "",
   "license": "Apache-2.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/modular-rocks/slimfast-turbo.git"
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org",
-    "access": "public"
-  },
-  "dependencies": {
-    "@babel/parser": "7.22.7",
-    "@babel/traverse": "7.22.8",
-    "@babel/types": "7.22.5",
-    "@modular-rocks/metrics-ts-js": "0.2.3-development",
-    "recast": "0.23.2"
   }
 }
