@@ -1,7 +1,4 @@
-// not sure why this is returning undefined when using import
-// works in some environments and not others
-// eslint-disable-next-line @typescript-eslint/no-require-imports
-const parser = require('@babel/parser');
+import { parse } from '@babel/parser';
 
 const babelConfig: RandomObject = {
   sourceType: 'module',
@@ -9,4 +6,4 @@ const babelConfig: RandomObject = {
   plugins: ['jsx', ['typescript', { isTSX: true }], 'babel-plugin-recast'],
 };
 
-export default (code: string) => parser.parse(code, babelConfig);
+export default (code: string) => parse(code, babelConfig);
